+++ src/__tests__/NetworkRequestInfo.spec.ts
@@ -5,6 +5,7 @@ jest.mock('react-native/Libraries/Blob/FileReader', () => ({}));
 describe('getCurlRequest', () => {
   it('should return valid curl request for simple GET request', () => {
     const info = new NetworkRequestInfo(
+      '1',
       'application/json',
       'GET',
       'https://test.com'
@@ -15,6 +16,7 @@ describe('getCurlRequest', () => {
 
   it('should return valid curl request for simple GET request with headers', () => {
     const info = new NetworkRequestInfo(
+      '1',
       'application/json',
       'GET',
       'https://test.com'
@@ -33,6 +35,7 @@ describe('getCurlRequest', () => {
 
   it('should return valid curl request for a POST request with headers', () => {
     const info = new NetworkRequestInfo(
+      '1',
       'application/json',
       'POST',
       'https://test.com/post'
@@ -49,6 +52,7 @@ describe('getCurlRequest', () => {
 
   it('should return valid curl request for a POST request with data', () => {
     const info = new NetworkRequestInfo(
+      '1',
       'application/json',
       'POST',
       'https://test.com/post'
@@ -67,6 +71,7 @@ describe('getCurlRequest', () => {
 
   it('should escape quotes in data', () => {
     const info = new NetworkRequestInfo(
+      '1',
       'application/json',
       'POST',
       'https://test.com/post'
@@ -90,11 +95,91 @@ describe('getCurlRequest', () => {
 
   it('should return valid curl request for a DELETE request', () => {
     const info = new NetworkRequestInfo(
+      '1',
       'application/json',
       'DELETE',
       'https://test.com'
     );
 
     expect(info.curlRequest).toEqual("curl -XDELETE 'https://test.com'");
+  });
+});
+
+describe('getRequestBody', () => {
+  const info = new NetworkRequestInfo(
+    '1',
+    'application/json',
+    'GET',
+    'https://test.com'
+  );
+
+  it('should return stringified data in consistent format', () => {
+    info.dataSent = '{"data":    {"a":   1   }}';
+    const result = info.getRequestBody();
+    expect(typeof result).toBe('string');
+    expect(result).toMatchInlineSnapshot(`
+      "{
+        \\"data\\": {
+          \\"a\\": 1
+        }
+      }"
+    `);
+  });
+
+  it('should return original object as string if stringify fails', () => {
+    // @ts-ignore
+    info.dataSent = { test: 1 };
+    const result = info.getRequestBody();
+    expect(typeof result).toBe('string');
+    expect(result).toEqual('[object Object]');
+  });
+
+  it('should process formData', () => {
+    const mockFormData = {
+      _parts: [
+        ['test', 'hello'],
+        ['another', 'goodbye'],
+      ],
+    };
+    // @ts-ignore
+    info.dataSent = mockFormData;
+    const result = info.getRequestBody();
+    expect(typeof result).toBe('string');
+    expect(result).toMatchInlineSnapshot(`
+      "{
+        \\"test\\": \\"hello\\",
+        \\"another\\": \\"goodbye\\"
+      }"
+    `);
+  });
+});
+
+describe('getResponseBody', () => {
+  const info = new NetworkRequestInfo(
+    '1',
+    'application/json',
+    'GET',
+    'https://test.com'
+  );
+
+  it('should return stringified data in consistent format', () => {
+    info.dataSent = '{"data":    {"a":   1   }}';
+    const result = info.getRequestBody();
+    expect(typeof result).toBe('string');
+    expect(result).toMatchInlineSnapshot(`
+      "{
+        \\"data\\": {
+          \\"a\\": 1
+        }
+      }"
+    `);
+  });
+
+  it('should return original object as string if stringify fails', () => {
+    // @ts-ignore
+    info.dataSent = { test: 1 };
+    const result = info.getRequestBody();
+    expect(typeof result).toBe('string');
+    expect(result).toEqual('[object Object]');
   });
 });